+++ src/MeetingsSDKAdapter.test.js
@@ -1,20 +1,22 @@
 import * as rxjs from 'rxjs';
-import {flatMap, take} from 'rxjs/operators';
+import {flatMap, take, last} from 'rxjs/operators';
 
 import MeetingSDKAdapter from './MeetingsSDKAdapter';
-import createMockSDK, {mockSDKMeeting} from './mockSdk';
+import createMockSDK from './mockSdk';
 
 describe('Meetings SDK Adapter', () => {
   let meeting;
   let meetingID;
   let meetingSDKAdapter;
   let mockSDK;
+  let mockSDKMeeting;
   let target;
 
   beforeEach(() => {
     mockSDK = createMockSDK();
-    meetingSDKAdapter = new MeetingSDKAdapter(mockSDK);
     meetingID = 'meetingID';
+    mockSDKMeeting = mockSDK.meetings.getMeetingByType('id', meetingID);
+    meetingSDKAdapter = new MeetingSDKAdapter(mockSDK);
     rxjs.fromEvent = jest.fn(() => rxjs.of({}));
     meeting = {
       ID: meetingID,
@@ -34,6 +36,7 @@ describe('Meetings SDK Adapter', () => {
     meeting = null;
     rxjs.fromEvent = null;
     mockSDK = null;
+    mockSDKMeeting = null;
     meetingSDKAdapter = null;
     meetingID = null;
     target = null;
@@ -340,7 +343,7 @@ describe('Meetings SDK Adapter', () => {
       meetingSDKAdapter.fetchMeetingTitle = jest.fn(() => Promise.resolve('my meeting'));
       meetingSDKAdapter.getLocalMedia = jest.fn(() => Promise.resolve({localAudio: 'localAudio', localVideo: 'localVideo'}));
 
-      meetingSDKAdapter.createMeeting(target).subscribe((newMeeting) => {
+      meetingSDKAdapter.createMeeting(target).pipe(last()).subscribe((newMeeting) => {
         expect(newMeeting).toMatchObject({...meeting, localAudio: 'localAudio', localVideo: 'localVideo'});
         done();
       });
@@ -352,8 +355,10 @@ describe('Meetings SDK Adapter', () => {
 
       meetingSDKAdapter.datasource.meetings.create = jest.fn(() => Promise.reject(errorMessage));
 
-      meetingSDKAdapter.createMeeting(target).subscribe(
-        () => {},
+      meetingSDKAdapter.createMeeting(wrongTarget).subscribe(
+        (meetingInstance) => {
+          done.fail('Unexpected message', meetingInstance);
+        },
         (error) => {
           expect(error).toBe(errorMessage);
           done();
@@ -428,8 +433,8 @@ describe('Meetings SDK Adapter', () => {
         expect(dataDisplay).toMatchObject({
           ID: 'mute-audio',
           icon: 'microphone-muted_28',
-          tooltip: 'Mute',
-          state: 'inactive',
+          tooltip: 'No microphone available',
+          state: 'disabled',
           text: null,
         });
         done();
@@ -489,7 +494,7 @@ describe('Meetings SDK Adapter', () => {
 
       expect(mockSDKMeeting.emit).toHaveBeenCalledWith('adapter:media:local:update', {
         control: 'mute-audio',
-        state: false,
+        state: 'inactive',
       });
     });
 
@@ -497,7 +502,10 @@ describe('Meetings SDK Adapter', () => {
       meetingSDKAdapter.meetings[meetingID].localAudio = null;
       await meetingSDKAdapter.handleLocalAudio(meetingID);
 
-      expect(mockSDKMeeting.unmuteAudio).toHaveBeenCalled();
+      expect(mockSDKMeeting.emit).toHaveBeenCalledWith('adapter:media:local:update', {
+        control: 'mute-audio',
+        state: 'disabled',
+      });
     });
 
     test('localAudio property should be defined once the audio track is unmuted', async () => {
@@ -514,7 +522,7 @@ describe('Meetings SDK Adapter', () => {
 
       expect(mockSDKMeeting.emit).toHaveBeenCalledWith('adapter:media:local:update', {
         control: 'mute-audio',
-        state: true,
+        state: 'disabled',
       });
     });
 
@@ -536,8 +544,8 @@ describe('Meetings SDK Adapter', () => {
         expect(dataDisplay).toMatchObject({
           ID: 'mute-video',
           icon: 'camera-muted_28',
-          tooltip: 'Stop video',
-          state: 'inactive',
+          tooltip: 'No camera available',
+          state: 'disabled',
           text: null,
         });
         done();
@@ -597,7 +605,7 @@ describe('Meetings SDK Adapter', () => {
 
       expect(mockSDKMeeting.emit).toHaveBeenCalledWith('adapter:media:local:update', {
         control: 'mute-video',
-        state: false,
+        state: 'inactive',
       });
     });
 
@@ -605,7 +613,10 @@ describe('Meetings SDK Adapter', () => {
       meetingSDKAdapter.meetings[meetingID].localVideo = null;
       await meetingSDKAdapter.handleLocalVideo(meetingID);
 
-      expect(mockSDKMeeting.unmuteVideo).toHaveBeenCalled();
+      expect(mockSDKMeeting.emit).toHaveBeenCalledWith('adapter:media:local:update', {
+        control: 'mute-video',
+        state: 'disabled',
+      });
     });
 
     test('localVideo property should be defined once the video track is unmuted', async () => {
@@ -622,7 +633,7 @@ describe('Meetings SDK Adapter', () => {
 
       expect(mockSDKMeeting.emit).toHaveBeenCalledWith('adapter:media:local:update', {
         control: 'mute-video',
-        state: true,
+        state: 'disabled',
       });
     });
 
@@ -641,6 +652,7 @@ describe('Meetings SDK Adapter', () => {
   describe('shareControl()', () => {
     test('returns the display data of a meeting control in a proper shape', (done) => {
       global.console.log = jest.fn();
+      meetingSDKAdapter.meetings[meetingID] = {...meeting};
 
       meetingSDKAdapter.shareControl(meetingID).subscribe((dataDisplay) => {
         expect(dataDisplay).toMatchObject({
@@ -656,10 +668,10 @@ describe('Meetings SDK Adapter', () => {
       global.console.log = jest.fn();
       meetingSDKAdapter.fetchMeeting = jest.fn();
 
-      meetingSDKAdapter.shareControl(meetingID).subscribe(
+      meetingSDKAdapter.shareControl('inexistent-meeting-id').subscribe(
         () => {},
         (error) => {
-          expect(error.message).toBe('Could not find meeting with ID "meetingID" to add share control');
+          expect(error.message).toBe('Could not find meeting with ID "inexistent-meeting-id" to add share control');
           done();
         },
       );
@@ -671,11 +683,6 @@ describe('Meetings SDK Adapter', () => {
     let stopStream;
 
     beforeEach(() => {
-      meetingSDKAdapter.meetings[meetingID] = {
-        ...meeting,
-        localShare: {},
-        remoteShare: {},
-      };
       const {trueStopStream} = meetingSDKAdapter.stopStream;
 
       stopStream = trueStopStream;
@@ -701,7 +708,7 @@ describe('Meetings SDK Adapter', () => {
     });
 
     test('start share if the share track is disabled', async () => {
-      meetingSDKAdapter.meetings[meetingID].localShare = null;
+      meetingSDKAdapter.meetings[meetingID] = {...meeting};
       const {getMediaStreams} = mockSDKMeeting;
 
       mockSDKMeeting.getMediaStreams = jest.fn(() => Promise.resolve([['mockStream'], 'localShare']));
@@ -713,14 +720,17 @@ describe('Meetings SDK Adapter', () => {
     });
 
     test('stop share if the share track is enabled', async () => {
-      meetingSDKAdapter.meetings[meetingID].localShare = 'localShare';
+      meetingSDKAdapter.meetings[meetingID] = {...meeting, localShare: 'localShare'};
       await meetingSDKAdapter.handleLocalShare(meetingID);
 
-      expect(mockSDKMeeting.updateShare).toHaveBeenCalled();
-      expect(meetingSDKAdapter.meetings[meetingID].localShare).toBeNull();
+      expect(mockSDKMeeting.updateShare).toHaveBeenCalledWith({
+        sendShare: false,
+        receiveShare: true,
+      });
     });
 
     test('resets sharing stream if share control is not handled properly', async () => {
+      meetingSDKAdapter.meetings[meetingID] = {...meeting, localShare: 'localShare'};
       global.console.warn = mockConsole;
       mockSDKMeeting.updateShare = jest.fn(() => Promise.reject());
       await meetingSDKAdapter.handleLocalShare(meetingID);
@@ -809,6 +819,7 @@ describe('Meetings SDK Adapter', () => {
       meetingSDKAdapter
         .createMeeting(target)
         .pipe(
+          last(),
           flatMap(() => meetingSDKAdapter.getMeeting(meetingID)),
           take(1),
         )